Extract Station Header components to their own files
