+++ app/javascript/react/components/molecules/FixedStreamStationHeader/atoms/DataSource/DataSource.style.tsx
@@ -1,0 +1,19 @@
+import styled from "styled-components";
+
+import media from "../../../../../utils/media";
+import { H4 } from "../../../../Typography";
+
+const DataDescription = styled(H4)`
+  display: inline;
+  font-weight: 500;
+  text-transform: uppercase;
+
+  @media ${media.desktop} {
+    display: block;
+    text-transform: none;
+    font-size: 18px;
+    font-weight: 400;
+  }
+`;
+
+export { DataDescription };
