+++ app/javascript/react/components/molecules/FixedStreamStationHeader/atoms/StationName/index.ts
@@ -1,0 +1,1 @@
+export * from "./StationName";
