+++ app/javascript/react/components/molecules/FixedStreamStationHeader/atoms/StreamUpdate/index.ts
@@ -1,0 +1,1 @@
+export * from "./StreamUpdate";
