+++ app/javascript/react/components/molecules/FixedStreamStationHeader/atoms/ValueLabel/index.ts
